% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colInfo.R
\name{colInfo}
\alias{colInfo}
\title{Information of a column in a data frame}
\usage{
colInfo(df, i)
}
\arguments{
\item{df}{Path to the input data frame}

\item{i}{The column name or index}
}
\value{
A data frame summarizing the information of the input data frame
}
\description{
This function summarizes some statistics
of a column in a data frame.
}
